refactor(itemsGrid): move onchain data processing logic to composable
